Move TimeHandle into a dedicated module
