+++ hydra-node/hydra-node.cabal
@@ -80,6 +80,7 @@ library
     Hydra.Chain.Direct
     Hydra.Chain.Direct.Context
     Hydra.Chain.Direct.Handlers
+    Hydra.Chain.Direct.TimeHandle
     Hydra.Chain.Direct.State
     Hydra.Chain.Direct.Tx
     Hydra.Chain.Direct.Util
